requirements.txt
# Core dependencies
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pydantic>=2.5.1
pydantic-settings>=2.1.0

# PDF Processing
PyPDF2>=3.0.1
pdf2image>=1.17.0
Pillow>=11.1.0
pikepdf>=8.13.0
img2pdf>=0.5.1
weasyprint>=60.2
reportlab>=4.0.8

# Data Processing
pandas>=2.1.3
numpy>=1.26.2
scikit-learn>=1.3.2

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiofiles>=23.2.1
python-magic>=0.4.27

# Testing
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1
